Support for global comparisons.
More methods added to support local and global estimate. If more than one
argument is passed through command line, the program creates objects and store
their values.

The goal is to be able to set speeds on a map, and compare different path
speeds between each other, setting a maximum and minimum globals and locals.